rename directories for service and json
