+++ web-server/src/server/app/app-effects.ts
@@ -8,9 +8,9 @@ import { Server, Socket } from 'socket.io';
 import { RoomEvent, ServerInfoEvent, SettingsEvent } from 'web-shared';
 import SocketService from '../clientio/socket-service';
 import CommandEffects from '../command/command-effects';
-import { CacheService } from '../eep/cache-service';
-import EepService from '../eep/eep-service';
-import EepDataEffects from '../json/eep-data-effects';
+import { CacheService } from '../eep-service/cache-service';
+import EepService from '../eep-service/eep-service';
+import EepDataEffects from '../eep-data/eep-data-effects';
 import LogEffects from '../log/log-effects';
 import AppConfig from './app-config';
 import AppReducer from './app-reducer';
