+++ web-server/src/server/eep-data/eep-data-reducer.ts
