+++ web-server/src/server/eep-data/observers/train/eep-rolling-stock-dto.ts
